requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
email-validator>=2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6 
PyJWT==2.10.1
requests==2.31.0
yfinance==0.2.50
